src/auth/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.1
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
PyJWT==2.6.0
PyMySQL==1.0.3
Werkzeug==2.3.1
